BitmexTracker/requirements.txt
bitmex-ws
elasticsearch
pandas
numpy